+++ packages/app/src/lib/app.ts
@@ -1,5 +1,0 @@
-export * from '@zerodep/can';
-export * from '@zerodep/errors';
-export * from '@zerodep/guard';
-export * from '@zerodep/is';
-export * from '@zerodep/to';