+++ packages/is/is/src/index.ts
@@ -6,6 +6,7 @@ export * from '@zerodep/is.error';
 export * from '@zerodep/is.float';
 export * from '@zerodep/is.function';
 export * from '@zerodep/is.integer';
+export * from '@zerodep/is.json';
 export * from '@zerodep/is.map';
 export * from '@zerodep/is.nil';
 export * from '@zerodep/is.null';